+++ packages/serverless-components/aws-cloudfront/__tests__/custom-url-origin.test.ts
@@ -44,7 +44,7 @@ describe("Input origin as a custom url", () => {
         maxTTL: 31536000,
         "lambda@edge": {
           "origin-request":
-            "arn:aws:lambda:us-east-1:123:function:originRequestFunction"
+            "arn:aws:lambda:us-west-2:123:function:originRequestFunction"
         }
       },
       origins: ["https://mycustomorigin.com"]