+++ packages/serverless-components/domain/src/component.ts
@@ -22,7 +22,7 @@ export default class Domain extends Component {
 
     this.context.debug("Validating inputs.");
 
-    inputs.region = inputs.region || "us-east-1";
+    inputs.region = inputs.region || "us-west-2";
     inputs.privateZone = inputs.privateZone || false;
     inputs.domainType = inputs.domainType || "both";
     inputs.defaultCloudfrontInputs = inputs.defaultCloudfrontInputs || {};